requirements.txt
# Core HTTP + async
httpx>=0.28.0
anyio>=4.0.0
httpcore>=1.0.0
h11>=0.16.0
sniffio>=1.3.0

# Telegram bot
python-telegram-bot>=20.0
pyee>=13.0.0

# Web scraping
requests>=2.32.0
beautifulsoup4>=4.14.0
soupsieve>=2.8.0

# Scheduling
APScheduler>=3.10.0
tzlocal>=5.0.0
tzdata>=2024.1

# Config
python-dotenv>=1.0.0

# Utilities
certifi>=2026.0.0
idna>=3.0
urllib3>=2.0.0
tqdm>=4.0.0
distro>=1.7.0
typing-extensions>=4.11.0
annotated-types>=0.7.0

# NOTE: playwright (for LinkedIn scraping) requires manual install:
#   python -m pip install playwright (then: playwright install chromium)
# NOTE: openai SDK replaced with direct httpx calls (no pydantic-core needed)
